New octree optimization:
I was checking all edge conditions, where one side of an object would be
inside of an edge of another octant. This however, was causing the tree
construction to take 3 times as long. As an alternative solution to edge
conditions, I s imply make the bounds of each octant larger by the
radius of the object, so that if an object is slightly inside of an
octant, it will be counted.